Giant commit (20 days of work).
Rigidbody bindings, major scripting improvements(prep for full serialization), compiles on windows, a lot of editor improvements (can now display a lot more things from C# scripts).

+++ Crowny/Source/Crowny/Scene/SceneManager.cpp
@@ -13,7 +13,7 @@ namespace Crowny
     void SceneManager::SetActiveScene(const Ref<Scene>& scene)
     {
         s_Scenes.push_back(scene);
-        s_ActiveIndex = s_Scenes.size() - 1;
+        s_ActiveIndex = (uint32_t)s_Scenes.size() - 1;
         ScriptRuntime::Init();
     }
