+++ Crowny/Source/Crowny/Scene/SceneRenderer.cpp
@@ -67,8 +67,8 @@ namespace Crowny
         {
             auto [transform, text] = scene->m_Registry.get<TransformComponent, TextComponent>(ee);
             Renderer2D::DrawString(text.Text, transform.GetTransform(), text.Font, text.Color);
-            s_Stats.Vertices += text.Text.size() * 6;
-            s_Stats.Triangles += text.Text.size() * 2;
+            s_Stats.Vertices += (uint32_t)text.Text.size() * 6;
+            s_Stats.Triangles += (uint32_t)text.Text.size() * 2;
         }
         Renderer2D::End();
         /*
